Implemented getCountries API call
+ CountryService
+ CountryRepository
+ New Migration to test the changes
+ Temp code in Program.cs to test the API call

Result -> each time the application is run, an getCountries is requested from the API which, through the new service, populates a Countries table with each country that Econt can operate with.

The way of populating the database from the API *IS* to be changed, as the rest of the calls would require more than 20-30 seconds (maybe even more) to fetch all the cities, offices, etc. from Econt's database. It would probably be changed to be done manually or by some script/service that calls it on a set timer form the day (when there would be little application traffic).

Rest of the Numenclature services (see EcontNumenclaturesService class) would be done analogically.