Database updating optimizations + bug fixes
Fixes:
- fixed a bug, where ServingOffices table did not populate properly (it's data wasn't fetched at all);
- fixed a bug, where populating cities into the database would update each city's country to the Country table, causing massive redundancy and duplication of data;
- fixed a bug, where REpopulating the database with data caused unnecessary deletions of old data. Now it properly first checks if there were any changes with the already fetched data before updating it.

The problem with the database updating being slow and memory consuming persists. But given that it needs to constantly either add or check for new instances of data (database records now going up to 400k), 1,5-1,6 GB of memory and ~45 seconds of wait time seems reasonable for now.

Further improvements are to be researched.